Add files via upload
The “Virtual Mouse Using Hand Gesture Recognition” project aims to provide a new and intuitive way of controlling a computer using hand gesture. And when our machine’s USB port not responding. At that time virtual mouse is very helpful to deal with that situation .
The combination of computer vision, Mediapipe and pyautogui provides a flexible and efficient solution for building and deploying the virtual mouse. Mediapipe’s hand landmark well in real world application and is a popular choice among researchers and developers in the field.